services/auth/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.2
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
loguru==0.7.3
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
win32_setctime==1.2.0
